+++ frontend/packages/frontend-customization/README.md
@@ -4,7 +4,7 @@ SPDX-FileCopyrightText: 2017-2020 City of Espoo
 SPDX-License-Identifier: LGPL-2.1-or-later
 -->
 
-# @evaka/evaka-customization
+# @evaka/frontend-customization
 
 Espoo implementation of interfaces for customizing eVaka frontend.
